+++ controllers/pkg/reconcilers/token/reconciler.go
@@ -20,10 +20,10 @@ import (
 	"context"
 	"fmt"
 	"reflect"
-	"strings"
 
 	"code.gitea.io/sdk/gitea"
 	"github.com/go-logr/logr"
+	commonv1alpha1 "github.com/nephio-project/api/common/v1alpha1"
 	infrav1alpha1 "github.com/nephio-project/api/infra/v1alpha1"
 	"github.com/nephio-project/nephio/controllers/pkg/giteaclient"
 	ctrlconfig "github.com/nephio-project/nephio/controllers/pkg/reconcilers/config"
@@ -33,6 +33,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/utils/pointer"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/log"
@@ -109,8 +110,10 @@ func (r *reconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Resu
 		// token being deleted
 		// Delete the token from the git server
 		// when successfull remove the finalizer
-		if err := r.deleteToken(ctx, giteaClient, cr); err != nil {
-			return ctrl.Result{Requeue: true}, errors.Wrap(r.Status().Update(ctx, cr), errUpdateStatus)
+		if cr.Spec.Lifecycle.DeletionPolicy == commonv1alpha1.DeletionDelete {
+			if err := r.deleteToken(ctx, giteaClient, cr); err != nil {
+				return ctrl.Result{Requeue: true}, errors.Wrap(r.Status().Update(ctx, cr), errUpdateStatus)
+			}
 		}
 
 		if err := r.finalizer.RemoveFinalizer(ctx, cr); err != nil {
@@ -172,7 +175,6 @@ func (r *reconciler) createToken(ctx context.Context, giteaClient *gitea.Client,
 			return err
 		}
 		r.l.Info("token created", "name", cr.GetName())
-		// owner reference dont work since this is a cross-namespace resource
 		secret := &corev1.Secret{
 			TypeMeta: metav1.TypeMeta{
 				APIVersion: corev1.SchemeGroupVersion.Identifier(),
@@ -182,6 +184,15 @@ func (r *reconciler) createToken(ctx context.Context, giteaClient *gitea.Client,
 				Namespace:   cr.GetNamespace(),
 				Name:        cr.GetName(),
 				Annotations: cr.GetAnnotations(),
+				OwnerReferences: []metav1.OwnerReference{
+					{
+						APIVersion: cr.APIVersion,
+						Kind:       cr.Kind,
+						Name:       cr.Name,
+						UID:        cr.UID,
+						Controller: pointer.Bool(true),
+					},
+				},
 			},
 			Data: map[string][]byte{
 				"username": []byte(u.UserName),
@@ -201,31 +212,11 @@ func (r *reconciler) createToken(ctx context.Context, giteaClient *gitea.Client,
 }
 
 func (r *reconciler) deleteToken(ctx context.Context, giteaClient *gitea.Client, cr *infrav1alpha1.Token) error {
-	secret := &corev1.Secret{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: corev1.SchemeGroupVersion.Identifier(),
-			Kind:       reflect.TypeOf(corev1.Secret{}).Name(),
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: cr.GetNamespace(),
-			Name:      cr.GetName(),
-		},
-	}
-	err := r.Delete(ctx, secret)
-	if resource.IgnoreNotFound(err) != nil {
-		r.l.Error(err, "cannot delete access token secret")
+	_, err := giteaClient.DeleteAccessToken(cr.GetTokenName())
+	if err != nil {
+		r.l.Error(err, "cannot delete token")
 		cr.SetConditions(infrav1alpha1.Failed(err.Error()))
 		return err
-	}
-
-	r.l.Info("token deleted", "name", cr.GetTokenName())
-	_, err = giteaClient.DeleteAccessToken(cr.GetTokenName())
-	if err != nil {
-		if !strings.Contains(err.Error(), "couldn't be found") {
-			r.l.Error(err, "cannot delete token")
-			cr.SetConditions(infrav1alpha1.Failed(err.Error()))
-			return err
-		}
 	}
 	r.l.Info("token deleted", "name", cr.GetTokenName())
 	return nil
